requirements.txt
paho-mqtt==1.1
pyserial==3.0.1
